+++ reference/html-proto/js/controller.js
@@ -2389,7 +2389,7 @@ function openCardPopup(card) {
         <div class="frame-cost">${vm.pipsHtml}</div>
       </div>
       <div class="frame-art">${vm.artInner}</div>
-      <div class="frame-type">${escapeHtml(vm.typeText)}</div>
+      <div class="frame-type">${vm.typeHtml}</div>
       <div class="frame-text">
         <div class="frame-oracle">${vm.oracleHtml}</div>
         ${vm.stickersInner ? '<div class="frame-stickers">' + vm.stickersInner + '</div>' : ''}